scripts/requirements.txt
librosa==0.10.1
parselmouth-praat==0.4.3
numpy==1.24.3
soundfile==0.12.1
